pipeline/requirements.txt
# Water-Watcher Scraping Pipeline
# Python 3.11+

# Core
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
alembic>=1.14,<2.0

# Scraping
beautifulsoup4>=4.12,<5.0
httpx>=0.28,<1.0
lxml>=5.0,<6.0
playwright>=1.49,<2.0

# Scheduling
apscheduler>=3.10,<4.0

# Data processing
pydantic>=2.10,<3.0
python-dateutil>=2.9,<3.0

# Notifications
pywebpush>=2.0,<3.0
py-vapid>=1.9,<2.0

# Config
python-dotenv>=1.0,<2.0

# Dev
pytest>=8.0,<9.0
pytest-asyncio>=0.25,<1.0
ruff>=0.9,<1.0
